requirements_backend.txt
fastapi
uvicorn
python-multipart
pydantic
sqlalchemy
passlib[bcrypt]
python-jose[cryptography]
psycopg2-binary
python-dotenv
google-generativeai
requests
